+++ tests/unit/__init__.py
@@ -1,0 +1,4 @@
+import os
+
+# setting up env for local
+os.environ["SPARKEXPECTATIONS_ENV"] = "local"
