+++ tests/unit/notification/plugins/test_base_notification.py
