+++ tests/unit/notification/plugins/test_email.py
