+++ tests/unit/notification/plugins/test_slack.py
