+++ tests/unit/sinks/plugins/test_base_writer.py
